feat(data): Migrate price/amount fields from f64 to Decimal
Resolves #48 by eliminating inconsistent numeric types across market data types.
All price and amount fields now use `rust_decimal::Decimal` for financial precision,
matching the precision pattern already established in `OrderBookL1` and rustrade-execution.

Changes:

Core data types:
  - `PublicTrade`: `price`, `amount` fields changed to `Decimal`
  - `Quote`: `bid_price`, `ask_price`, `bid_amount`, `ask_amount` fields changed to `Decimal`
  - `Candle`: `open`, `high`, `low`, `close`, `volume` fields changed to `Decimal`
  - `Liquidation`: `price`, `quantity` fields changed to `Decimal`
  - `Quote` made `Copy` for better ergonomics with the smaller `Decimal` footprint

Quote utilities refactored:
  - `mid_price()` returns `Decimal` instead of `f64`
  - `spread()` returns `Decimal` instead of `f64`
  - `spread_bps()` returns `Decimal`, uses constant `BPS_FACTOR` instead of inline `10_000.0`
  - Comparison logic updated to use `Decimal::ZERO` and `.is_zero()` instead of float comparisons

Exchange transformers updated:
  - Alpaca (equities and crypto trades)
  - Binance (futures liquidations and trades)
  - Bitfinex, BitMEX, Bybit, Coinbase, Databento, GateIO, Hyperliquid, IBKR, Kraken, OKX
  - All now deserialize and transform prices/amounts as `Decimal`

Tests refactored:
  - Updated fixtures to use `dec!()` macro for precision
  - Fixed Quote::spread_bps test to use range-based comparison for Decimal precision
  - Updated all assertion values to match Decimal semantics

Documentation:
  - CHANGELOG.md: Added migration guide and breaking change details
  - Documented conversion patterns: `dec!()` for literals, `.parse::<Decimal>()` for strings,
    `Decimal::try_from(f64)` only when source is already `f64`

+++ rustrade-data/src/exchange/bitfinex/trade.rs
@@ -3,6 +3,7 @@ use crate::{
     subscription::trade::PublicTrade,
 };
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 use rustrade_instrument::{Side, exchange::ExchangeId};
 use rustrade_integration::serde::de::{datetime_utc_from_epoch_duration, extract_next};
 use serde::Serialize;
@@ -39,8 +40,8 @@ pub struct BitfinexTrade {
     pub id: u64,
     pub time: DateTime<Utc>,
     pub side: Side,
-    pub price: f64,
-    pub amount: f64,
+    pub price: Decimal,
+    pub amount: Decimal,
 }
 
 impl<InstrumentKey> From<(ExchangeId, InstrumentKey, BitfinexTrade)>
@@ -86,7 +87,7 @@ impl<'de> serde::Deserialize<'de> for BitfinexTrade {
                 // Trade: [ID, TIME, AMOUNT,PRICE]
                 let id = extract_next(&mut seq, "id")?;
                 let time_millis = extract_next(&mut seq, "time")?;
-                let amount: f64 = extract_next(&mut seq, "amount")?;
+                let amount: Decimal = extract_next(&mut seq, "amount")?;
                 let price = extract_next(&mut seq, "price")?;
                 let side = match amount.is_sign_positive() {
                     true => Side::Buy,
